Fix multiple compiler, SWIG, and wrapper redefinition warnings
- src/cbf_hdf5.c: Resolved -Wstringop-overflow by replacing potentially
  underflowing loops with safer while(ii > 0) patterns, and fixed
  -Wformat-overflow by adding NULL checks for 'ivalue' before sprintf.
- src/cbf_uncompressed.c: Fixed _XOPEN_SOURCE redefinition by moving
  it before system includes.
- src/jcbf.i: Resolved SWIG Warning 451 by marking cbf_node_struct::name
  as %immutable.
- pycbf/: Fixed SWIG_PYTHON_STRICT_BYTE_CHAR redefinitions using #ifndef.
- examples/adscimg2cbf_sub.c: Fixed -Wformat-truncation using snprintf
  and increasing 'temp' buffer size.
- examples/cbf_standardize_numbers.c: Fixed -Wformat-overflow by
  increasing 'format' buffer size.
- examples/changtestcompression.c: Fixed -Wimplicit-const-int-float-conversion
  using explicit casts for large constants.

Verified with clean builds and testhdf5 suite. No binary artifacts included.

+++ examples/changtestcompression.c
@@ -119,7 +119,7 @@ static size_t createtestimage(void **data, int type, int nelem)
     uldata = (unsigned long *)calloc(nelem, sizeof(long));
     elsize = sizeof(long);
     while (i < nelem) {
-      uldata[i] = urand() * ULONG_MAX;
+      uldata[i] = (unsigned long)(urand() * (double)ULONG_MAX);
       i += 1 + urand() * NSTEPS;
     }
     *data = uldata;
@@ -128,7 +128,7 @@ static size_t createtestimage(void **data, int type, int nelem)
     sldata = (signed long *)calloc(nelem, sizeof(long));
     elsize = sizeof(long);
     while (i < nelem) {
-      sldata[i] = urand() * LONG_MAX;
+      sldata[i] = (signed long)(urand() * (double)LONG_MAX);
       i += 1 + urand() * NSTEPS;
     }
     *data = sldata;
@@ -138,7 +138,7 @@ static size_t createtestimage(void **data, int type, int nelem)
     ulldata = (unsigned long long *)calloc(nelem, sizeof(long long));
     elsize = sizeof(long long);
     while (i < nelem) {
-      ulldata[i] = urand() * ULLONG_MAX;
+      ulldata[i] = (unsigned long long)(urand() * (double)ULLONG_MAX);
       i += 1 + urand() * NSTEPS;
     }
     *data = ulldata;
@@ -147,7 +147,7 @@ static size_t createtestimage(void **data, int type, int nelem)
     slldata = (signed long long *)calloc(nelem, sizeof(long long));
     elsize = sizeof(long long);
     while (i < nelem) {
-      slldata[i] = urand() * LLONG_MAX;
+      slldata[i] = (signed long long)(urand() * (double)LLONG_MAX);
       i += 1 + urand() * NSTEPS;
     }
     *data = slldata;
